requirements.txt
fastapi
uvicorn[standard]
jinja2
asyncpg
ormar
psycopg2-binary
pydantic
python-multipart
typing
python-jose